move abstract parts from MapDownloadJob to DownloadJob class
